refactor: move utility functions to their own modules, extract static calculate entropy method to data_util module
